requirements.txt
PyQt5>=5.15.4
selenium-wire>=4.5.1
slackclient>=2.9.3